[consensus] moving pending_votes out of block_store
